ci: 关掉 clang-tidy 21+ 新检查 + 修 conan profile 在 GCC 16 下的报错
切到 archlinux:base-devel 后 clang-tidy 升到 22 / GCC 升到 16 引入两类新失败：

1) clang-tidy 21/22 新增 / 升 warning 级的检查，触发 master 历史代码：
   - cppcoreguidelines-pro-bounds-avoid-unchecked-container-access (v[i])
   - modernize-use-designated-initializers / modernize-use-auto
   - bugprone-throwing-static-initialization / bugprone-exception-escape
   - bugprone-derived-method-shadowing-base-method
   - readability-use-concise-preprocessor-directives
   .clang-tidy 显式禁用这一批，保持检查面与原 clang-tidy 20 一致。
   想开任何一条，要先把对应代码模式（v[i]、聚合 init、static throw 等）
   逐处治理过再开 —— 那不是本 PR 的范围。

2) conan profile 在 GCC 16 下的 'Invalid setting' 报错：
   conan 2.28.1 settings.yml 最高承认 15.2，而 Arch 滚动 GCC 已到 16。
   `conan profile detect --force` 把 default profile 写成 16，install 时
   --profile:build 解析失败。改成 --profile:all 把 host + build profile
   都指向 vendored conan/profiles/linux-gcc（compiler.version=15），
   彻底绕开 default profile。等 conan 后续 release 认 GCC 16 再改回去。

附带（lint 仍命中、需要按个修的几处）：
- placement_new.cpp: WithConst 上的 NOLINTNEXTLINE 错位，移到 `int const id;` 行上方
- multiple_inheritance.cpp: StringHolder(int n) 显式 size_t cast，避免
  clang-diagnostic-sign-conversion

+++ modules/02_lifetime_type_safety/demos/multiple_inheritance.cpp
@@ -7,6 +7,7 @@
 //   - 构造函数继承（using Base::Base）允许派生类直接复用父类构造函数 —— 但
 //     编译器自动生成的特殊成员（如默认构造）不会被继承。
 
+#include <cstddef>
 #include <iostream>
 #include <string>
 #include <utility>
@@ -44,7 +45,7 @@ struct Triathlete : Swimmer, Runner {
 struct StringHolder {
     std::string value;
     explicit StringHolder(std::string v) : value(std::move(v)) {}
-    explicit StringHolder(int n) : value(n, 'x') {}  // n 个 'x'
+    explicit StringHolder(int n) : value(static_cast<std::size_t>(n), 'x') {}  // n 个 'x'
     void show() const {
         std::cout << "  value=" << value << "\n";
     }
